# docs/openapi/openapi.yaml
components:
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  contact: {}
  description: Deterministic, highly resilient backend infrastructure for the Lemni Payment-as-a-Service (PaaS) engine.
  title: LEMNI PaaS Core Engine API
  version: 1.0.0
openapi: 3.0.0
paths:
...
  /api/v1/checkout/plans/{planId}/sessions:
    post:
      description: Allows off-the-shelf payment links to initialize checkout sessions for customer emails.
      operationId: CheckoutController_createPublicPlanSession
      parameters:
        - description: The static pricing plan ID
          in: path
          name: planId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                callbackUrl:
                  example: https://mywebsite.com/public-success
                  format: uri
                  type: string
                email:
                  example: buyer@gmail.com
                  format: email
                  type: string
              required:
                - email
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  checkoutUrl:
                    description: URL to direct customer to for payment
                    example: https://checkout.nomba.com/pay/mock_link_123
                    format: uri
                    type: string
                  sessionId:
                    description: Unique session ID for tracking
                    example: tx_1234567890abcdef
                    type: string
                  subscriptionId:
                    description: Subscription created for this session
                    example: sub_a28deca9
                    type: string
                type: object
          description: Checkout session generated successfully
        '400':
          content:
            application/json:
              schema:
                properties:
                  error:
                    example: Invalid email format or plan configuration
                    type: string
                type: object
          description: Invalid input
        '404':
          content:
            application/json:
              schema:
                properties:
                  error:
                    example: Plan not found
                    type: string
                type: object
          description: Plan not found
      summary: Generate public plan checkout session
      tags:
        - developer-apis/checkout
  /api/v1/pay:
    post:
      description: Generates a payment URL for a singular checkout charge.
      operationId: CheckoutController_createOneTimePayment
      parameters:
        - description: Bearer <API_KEY>
          in: header
          name: Authorization
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                amount:
                  example: 5000
                  format: float
                  type: number
                callbackUrl:
                  example: https://mywebsite.com/payment-success
                  format: uri
                  type: string
                email:
                  example: payer@test.com
                  format: email
                  type: string
              required:
                - amount
                - email
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  checkoutUrl:
                    description: URL to direct customer to for payment
                    example: https://checkout.nomba.com/pay/mock_link_123
                    format: uri
                    type: string
                  sessionId:
                    description: Unique transaction/session ID for tracking and polling status
                    example: tx_1234567890abcdef
                    type: string
                type: object
          description: Checkout session successfully created
        '400':
          content:
            application/json:
              schema:
                properties:
                  error:
                    example: Invalid amount or email format
                    type: string
                type: object
          description: Invalid input
        '401':
          content:
            application/json:
              schema:
                properties:
                  error:
                    example: Unauthorized API key
                    type: string
                type: object
          description: Unauthorized API key
      summary: Create a one-time checkout session
      tags:
        - developer-apis/checkout
  /api/v1/public/subscriptions/{id}/unsubscribe/confirm:
    post:
      description: Verifies the OTP code sent to the subscription owner email and cancels the subscription immediately.
      operationId: CheckoutController_confirmUnsubscribe
      parameters:
        - description: The subscription ID
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                code:
                  description: 6-digit OTP code sent to email
                  example: '123456'
                  type: string
              required:
                - code
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  message:
                    description: Confirmation message
                    example: Subscription successfully canceled
                    type: string
                  success:
                    description: Success flag
                    example: true
                    type: boolean
                type: object
          description: Subscription successfully canceled
        '400':
          content:
            application/json:
              schema:
                properties:
                  error:
                    description: Error message
                    example: Invalid or expired code, or subscription already canceled
                    type: string
                type: object
          description: Invalid or expired code, or subscription already canceled
        '404':
          content:
            application/json:
              schema:
                properties:
                  error:
                    description: Error message
                    example: Subscription not found
                    type: string
                type: object
          description: Subscription not found
      summary: Confirm unsubscribe with OTP
      tags:
        - developer-apis/checkout
  /api/v1/public/subscriptions/{id}/unsubscribe/request:
    post:
      description: Generates and sends a 6-digit OTP code to the subscription owner email to verify unsubscribe request.
      operationId: CheckoutController_requestUnsubscribe
      parameters:
        - description: The subscription ID
          in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                email:
                  description: Email address associated with the subscription
                  example: subscriber@test.com
                  format: email
                  type: string
              required:
                - email
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  message:
                    description: Confirmation message
                    example: Verification code sent to subscriber@test.com
                    type: string
                type: object
          description: Verification code sent successfully
        '403':
          content:
            application/json:
              schema:
                properties:
                  error:
                    description: Error message
                    example: Email does not match subscription owner
                    type: string
                type: object
          description: Email does not match subscription owner
        '404':
          content:
            application/json:
              schema:
                properties:
                  error:
                    description: Error message
                    example: Subscription not found
                    type: string
                type: object
          description: Subscription not found
      summary: Request email OTP to unsubscribe
      tags:
        - developer-apis/checkout
  /api/v1/sessions/{id}/status:
    get:
      description: Enables frontend checkouts or backend applications to poll for session completion status.
      operationId: CheckoutController_getCheckoutSessionStatus
      parameters:
        - description: The unique session ID / transaction ID
          in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  amount:
                    description: Transaction amount
                    example: 5000
                    format: float
                    type: number
                  createdAt:
                    description: Session creation timestamp
                    example: '2026-07-04T02:00:00.000Z'
                    format: date-time
                    type: string
                  nombaRef:
                    description: Reference from payment processor (null until payment is processed)
                    example: ref_nomba_992c
                    nullable: true
                    type: string
                  sessionId:
                    description: The session/transaction ID
                    example: tx_1234567890abcdef
                    type: string
                  status:
                    description: Current payment status
                    enum:
                      - pending
                      - success
                      - failed
                    example: pending
                    type: string
                type: object
          description: Session status retrieved successfully
        '404':
          content:
            application/json:
              schema:
                properties:
                  error:
                    example: Session not found
                    type: string
                type: object
          description: Session not found
      summary: Poll checkout session status
      tags:
        - developer-apis/checkout
  /api/v1/subscribe:
    post:
      description: Generates a payment URL to register a customer to a recurring pricing plan.
      operationId: CheckoutController_createSubscriptionPayment
      parameters:
        - description: Bearer <API_KEY>
          in: header
          name: Authorization
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                callbackUrl:
                  example: https://mywebsite.com/sub-success
                  format: uri
                  type: string
                email:
                  example: subscriber@test.com
                  format: email
                  type: string
                planId:
                  example: plan_7a8dca9
                  type: string
              required:
                - planId
                - email
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  checkoutUrl:
                    description: URL to direct customer to for payment
                    example: https://checkout.nomba.com/pay/mock_link_123
                    format: uri
                    type: string
                  sessionId:
                    description: Unique transaction/session ID
                    example: tx_1234567890abcdef
                    type: string
                  subscriptionId:
                    description: Subscription created for this session
                    example: sub_a28deca9
                    type: string
                type: object
          description: Checkout session successfully created
        '400':
          content:
            application/json:
              schema:
                properties:
                  error:
                    example: Invalid plan ID or email format
                    type: string
                type: object
          description: Invalid input
        '401':
          content:
            application/json:
              schema:
                properties:
                  error:
                    example: Unauthorized API key
                    type: string
                type: object
          description: Unauthorized API key
        '404':
          content:
            application/json:
              schema:
                properties:
                  error:
                    example: Plan not found
                    type: string
                type: object
          description: Plan not found
      summary: Create a subscription checkout session
      tags:
        - developer-apis/checkout
...
servers: []
...